requirements.txt
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-SQLAlchemy==2.3.2
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==4.1.2
packaging==17.1
passlib==1.7.1
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyparsing==2.2.0
python-decouple==3.1
python-dotenv==0.9.1
redis==2.10.6
requests==2.19.1
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
virtualenv==16.0.0
webencodings==0.5.1
Werkzeug==0.14.1
